cleanup2

+++ Quol-CPP/src/core/PluginStoreManager.cpp
@@ -13,6 +13,8 @@
 #include <QTemporaryFile>
 #include <QUrl>
 
+#include <functional>
+
 static const char *kStoreApiUrl = "https://api.github.com/repos/LeoCh01/Quol-Tools/contents/plugins?ref=main";
 static const char *kRawBaseUrl = "https://raw.githubusercontent.com/LeoCh01/Quol-Tools/main/plugins/";
 
@@ -41,6 +43,15 @@ void PluginStoreManager::fetchStoreItems() {
     });
 }
 
+QString PluginStoreManager::artifactPluginName(const QString &itemName) {
+    QString base = itemName;
+    if (base.endsWith(QStringLiteral(".zip"))) {
+        base.chop(4);
+    }
+    const int sep = base.lastIndexOf(QStringLiteral("--v"));
+    return (sep != -1) ? base.left(sep) : base;
+}
+
 int PluginStoreManager::compareVersions(const QString &a, const QString &b) {
     const QStringList aParts = a.split('.');
     const QStringList bParts = b.split('.');
@@ -56,6 +67,34 @@ int PluginStoreManager::compareVersions(const QString &a, const QString &b) {
     return 0;
 }
 
+void PluginStoreManager::extractZipAsync(
+    const QString &zipPath, const QString &pluginDir, std::function<void(bool)> onDone
+) {
+    const QString script = QStringLiteral("Expand-Archive -LiteralPath \"%1\" -DestinationPath \"%2\" -Force")
+                               .arg(QDir::toNativeSeparators(zipPath), QDir::toNativeSeparators(pluginDir));
+
+    auto *proc = new QProcess(this);
+    connect(
+        proc,
+        qOverload<int, QProcess::ExitStatus>(&QProcess::finished),
+        this,
+        [this, proc, zipPath, pluginDir, onDone](int exitCode, QProcess::ExitStatus) {
+            proc->deleteLater();
+            QFile::remove(zipPath);
+
+            const QString pluginName = QFileInfo(pluginDir).fileName();
+            const QString configPath = pluginDir + QStringLiteral("/res/config.json");
+            const QString dllPath = pluginDir + QStringLiteral("/") + pluginName + QStringLiteral(".dll");
+
+            onDone(exitCode == 0 && QFile::exists(configPath) && QFile::exists(dllPath));
+        }
+    );
+    proc->start(
+        QStringLiteral("powershell"),
+        {QStringLiteral("-NoProfile"), QStringLiteral("-NonInteractive"), QStringLiteral("-Command"), script}
+    );
+}
+
 void PluginStoreManager::downloadPlugin(const QString &itemName, bool isUpdate, const QString &appVersion) {
     const QString url = QLatin1String(kRawBaseUrl) + itemName + QStringLiteral(".zip");
     QNetworkRequest request;
@@ -66,11 +105,7 @@ void PluginStoreManager::downloadPlugin(const QString &itemName, bool isUpdate, 
     connect(reply, &QNetworkReply::finished, this, [this, reply, itemName, isUpdate, appVersion]() {
         reply->deleteLater();
 
-        // Derive folder name by stripping "--v<version>" suffix
-        QString pluginName = itemName;
-        const int sep = pluginName.lastIndexOf(QStringLiteral("--v"));
-        if (sep != -1)
-            pluginName = pluginName.left(sep);
+        const QString pluginName = artifactPluginName(itemName);
 
         if (reply->error() != QNetworkReply::NoError) {
             emit pluginDownloadFinished(pluginName, false);
@@ -81,8 +116,6 @@ void PluginStoreManager::downloadPlugin(const QString &itemName, bool isUpdate, 
         const QString pluginsDir = QCoreApplication::applicationDirPath() + QStringLiteral("/plugins");
         const QString pluginDir = pluginsDir + QStringLiteral("/") + pluginName;
         const QString backupDir = pluginsDir + QStringLiteral("/") + pluginName + QStringLiteral("_backup");
-        const QString configPath = pluginDir + QStringLiteral("/res/config.json");
-        const QString dllPath = pluginDir + QStringLiteral("/") + pluginName + QStringLiteral(".dll");
 
         // Save downloaded zip to a temp file
         QTemporaryFile tempZip(QDir::tempPath() + QStringLiteral("/quol_plugin_XXXXXX.zip"));
@@ -108,26 +141,12 @@ void PluginStoreManager::downloadPlugin(const QString &itemName, bool isUpdate, 
             return;
         }
 
-        // Use PowerShell Expand-Archive for zip extraction (async, non-blocking)
-        const QString script = QStringLiteral("Expand-Archive -LiteralPath \"%1\" -DestinationPath \"%2\" -Force")
-                                   .arg(QDir::toNativeSeparators(zipPath), QDir::toNativeSeparators(pluginDir));
-
-        auto *proc = new QProcess(this);
-        connect(
-            proc,
-            qOverload<int, QProcess::ExitStatus>(&QProcess::finished),
-            this,
-            [this, proc, pluginName, isUpdate, zipPath, pluginDir, backupDir, appVersion](
-                int exitCode, QProcess::ExitStatus
-            ) {
-                proc->deleteLater();
-                QFile::remove(zipPath);
-
-                const QString configPath = pluginDir + QStringLiteral("/res/config.json");
-                const QString dllPath = pluginDir + QStringLiteral("/") + pluginName + QStringLiteral(".dll");
-                bool ok = exitCode == 0 && QFile::exists(configPath) && QFile::exists(dllPath);
+        extractZipAsync(
+            zipPath, pluginDir, [this, pluginName, isUpdate, pluginDir, backupDir, appVersion](bool extracted) {
+                bool ok = extracted;
 
                 if (ok && !appVersion.isEmpty()) {
+                    const QString configPath = pluginDir + QStringLiteral("/res/config.json");
                     QFile cf(configPath);
                     if (cf.open(QIODevice::ReadOnly | QIODevice::Text)) {
                         const QJsonDocument doc = QJsonDocument::fromJson(cf.readAll());
@@ -151,29 +170,18 @@ void PluginStoreManager::downloadPlugin(const QString &itemName, bool isUpdate, 
                     } else if (QDir(pluginDir).exists()) {
                         QDir(pluginDir).removeRecursively();
                     }
-                } else {
-                    if (isUpdate && QDir(backupDir).exists())
-                        QDir(backupDir).removeRecursively();
+                } else if (isUpdate && QDir(backupDir).exists()) {
+                    QDir(backupDir).removeRecursively();
                 }
 
                 emit pluginDownloadFinished(pluginName, ok);
             }
         );
-        proc->start(
-            QStringLiteral("powershell"),
-            {QStringLiteral("-NoProfile"), QStringLiteral("-NonInteractive"), QStringLiteral("-Command"), script}
-        );
     });
 }
 
 void PluginStoreManager::installLocalPlugin(const QString &zipFilePath) {
-    QFileInfo fi(zipFilePath);
-    QString baseName = fi.completeBaseName();
-
-    QString pluginName = baseName;
-    const int sep = pluginName.lastIndexOf(QStringLiteral("--v"));
-    if (sep != -1)
-        pluginName = pluginName.left(sep);
+    const QString pluginName = artifactPluginName(QFileInfo(zipFilePath).completeBaseName());
 
     const QString pluginsDir = QCoreApplication::applicationDirPath() + QStringLiteral("/plugins");
     const QString pluginDir = pluginsDir + QStringLiteral("/") + pluginName;
@@ -207,32 +215,10 @@ void PluginStoreManager::installLocalPlugin(const QString &zipFilePath) {
     srcFile.close();
     const QString zipPath = tempZip.fileName();
 
-    const QString script = QStringLiteral("Expand-Archive -LiteralPath \"%1\" -DestinationPath \"%2\" -Force")
-                               .arg(QDir::toNativeSeparators(zipPath), QDir::toNativeSeparators(pluginDir));
-
-    auto *proc = new QProcess(this);
-    connect(
-        proc,
-        qOverload<int, QProcess::ExitStatus>(&QProcess::finished),
-        this,
-        [this, proc, pluginName, pluginDir, zipPath](int exitCode, QProcess::ExitStatus) {
-            proc->deleteLater();
-            QFile::remove(zipPath);
-
-            const QString configPath = pluginDir + QStringLiteral("/res/config.json");
-            const QString dllPath = pluginDir + QStringLiteral("/") + pluginName + QStringLiteral(".dll");
-            bool ok = exitCode == 0 && QFile::exists(configPath) && QFile::exists(dllPath);
-
-            if (!ok) {
-                if (QDir(pluginDir).exists())
-                    QDir(pluginDir).removeRecursively();
-            }
-
-            emit localPluginInstallFinished(pluginName, ok);
+    extractZipAsync(zipPath, pluginDir, [this, pluginName, pluginDir](bool ok) {
+        if (!ok && QDir(pluginDir).exists()) {
+            QDir(pluginDir).removeRecursively();
         }
-    );
-    proc->start(
-        QStringLiteral("powershell"),
-        {QStringLiteral("-NoProfile"), QStringLiteral("-NonInteractive"), QStringLiteral("-Command"), script}
-    );
+        emit localPluginInstallFinished(pluginName, ok);
+    });
 }
